Split code in separate files mimicking the structure of solvers package
