+++ pysmt/cmd/installers/__init__.py
@@ -1,0 +1,7 @@
+from pysmt.cmd.installers.msat import MSatInstaller
+from pysmt.cmd.installers.z3 import Z3Installer
+from pysmt.cmd.installers.cvc4 import CVC4Installer
+from pysmt.cmd.installers.yices import YicesInstaller
+from pysmt.cmd.installers.btor import BtorInstaller
+from pysmt.cmd.installers.pico import PicoSATInstaller
+from pysmt.cmd.installers.bdd import CuddInstaller
